+++ .vscode-test.mjs
@@ -1,5 +1,9 @@
 import { defineConfig } from '@vscode/test-cli'
 
 export default defineConfig({
-  files: 'out/test/**/*.test.js'
+  files: 'out/test/**/*.test.js',
+  workspaceFolder: './src/test/intellisense',
+  mocha: {
+    timeout: 20_000
+  }
 })