+++ src/test/intellisense/Lib/AppGlobalScript.js
@@ -1,0 +1,6 @@
+/**
+ * @module global module
+ * all symbols in global scope available in whole project
+ */
+
+const USER = 'Maxim'