+++ src/test/intellisense/dep.js
@@ -1,3 +1,9 @@
+/**
+ * Sum two numbers
+ * @param {number} a first value
+ * @param {number} b second value
+ * @returns {number}
+ */
 function sum(a, b) {
   return a + b
 }