+++ src/test/intellisense/entry.js
@@ -1,3 +1,7 @@
 import 'dep.js'
 
 sum(1, 2)
+
+if (USER === 'Maxim') {
+  alert(IDOK)
+}